ci: Factor out a script to install a clear image
Just like the Clear kernel, we ought to be able to install a clear image
with one command line.

I took the opportunity to make this more general. We can install any
version now instead of the just the latest one. The decision of which
version to install is still left to the "master script",
install_cc_env_ubuntu.sh

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>