Makefile: Re-enable and fix CRI-O tests on Semaphore CI
This patch re-enables the CRI-O tests because they are now fixed by
setting the environment variable "STORAGE_OPTS". That way, we can
specify which storage driver we want to use.

In case of Semaphore, because it is running Ubuntu 14.04 on a 3.13
kernel, we have to set aufs as the storage driver, otherwise we get
some issues using devicemapper.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>