checkcommits: Check first character after looking for "Fixes" comment.
The first character in each line is special: if it is non-alphabetic,
line limits don't apply. However, previously that check was performed
before the "Fixes #XXX" check meaning if a "Fixes #XXX" line was
indented, it would not be considered. This would result in a failure if
the user specified "--need-fixes".

With this change, an indented "Fixes #XXX" is now permitted.

Fixes #113.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>