checkcommits: Create regexp patterns in constructor.
Move the regexp pattern compilation to NewCommitConfig(). This makes the
code simpler and safer: previously, the tests were creating their own
patterns which were not the same as the tool (the tests were allowing a
case insensitive "Fixes #XXX" match, whereas the tool was only allowing
a leading-caps match).

This fix also restricts performing the regexp matching to those cases
where the user requires a SOB or Fixes comment.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>