functional: Implement standalone testing of the exec command
We are submitting patches to the runtime repository in order to
fix the behavior of the exec command, but we were missing some
tests to validate those expected behaviors.

This patch adds some testing of the exec command when it is
either started "attached" or "detached" from the shell.

Fixes #171

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>