.ci: Install specific vmlinux and vmlinuz versions for Semaphore CI
We know that we have a limitation with Semaphore, we have to stick
to one specific kernel version. Because we are trying to move to
the "pc" machine type, we need a compressed kernel, but it is not
provided by the Clear Linux server for this specific version.
Indeed, the vmlinuz has been added recently to our packaging, and
for that reason, it does not exist for 4.5-50 version.

Both vmlinux and vmlinuz have been re-generated from sources and
since we know that we won't change those binaries, there is no
point in trying to download and extract them from Clear Linux
server. This saves some time for our CI.

This patch first adds both kernel binaries to the repository, and
then it uses them from the CI scripts so that we copy and link
them as the expected vmlinux.container and vmlinuz.container.

Fixes #350

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>