ci: Increase test timeout from 5 to 15 seconds
The 5 seconds timeout is appropriate if the tests run on baremetal,
but according to recent changes on QEMU parameters if running in a
nested environment, the tests have been slowed down by disabling fast
MMIO access previously performed by virtio 1.0.

And given the fact that our CI runs on a VMM, we cannot afford this
timeout of 5 seconds. Moving this timeout to 15 seconds makes the
tests passing even when run inside a VM.

Fixes #412

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>