metrics: ci: Add more tests to the metrics ci script
We don't run all the metrics tests for the metrics CI, but we do
want to run a 'good subset' that covers all key data from all key
areas of interest.
Thus, add memory footprint, network and storage tests to the default
set of metrics we run in the ci.

Fixes: #414

Signed-off-by: Graham Whaley <graham.whaley@intel.com>