cmd: emailreport: Notify errors by email
This commit allows that despite checkmetrics tool fails,
emailreport tool sends a notification about that, also
this commit includes the support of “comments” option,
the idea is to send an email notification about checkmetrics
results and using the option comments, it can make
reference to a specific event.
E.g. a pull request.

Signed-off-by: Mario Alfredo Carrillo Arevalo <mario.alfredo.c.arevalo@intel.com>