kubernetes: pull images before launching workload
The nginx test has a timeout for launching the workload, as
the required images are downloaded at that time, sometimes the
test fails with a timeout. To prevent this, this fix pulls the
image before launching the workload.

Also when initializing the kubernetes services, including the
dns pod, sometimes it is not enough 60 seconds to pull all images
and start the services, then, this also increases the timeout of
the initialization to 180 secs.

Fixes #511.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>