metrics: network: Simplify latency test
This commit removes redundant code and it adapts
function according to the new changes in network lib.
furthermore it was tested using :runc" as runtime.

Another change is that this time the server container
will finish in the end of the test, and it is not
conditioned to a sleep time.

Signed-off-by: Mario Alfredo Carrillo Arevalo <mario.alfredo.c.arevalo@intel.com>