ci: install_proxy: disable proxy ksm mode if on metrics CI
When running under the metrics CI system we do not want the proxy
auto-modifying the KSM modes - we control those from the CI to ensure
we know exactly what we are testing and to get steady state results.

Modify the .ci proxy install script so that it adds the correct
argument to the sytemd file, or copies the correct (added in this
PR) upstart file into place.

Also, add the '-log debug' flag into the systemd file, which was an
existing oversight (it was enabled in the upstart file).

Fixes: #559

Signed-off-by: Graham Whaley <graham.whaley@intel.com>