makefile: Remove legacy AUFS testing for CRI-O
The purpose of forcing the storage driver to be AUFS was done because
our previous CI (Semaphore) was relying on Ubuntu 14.04 as OS and this
was the only way to have stable testing.

Now that we have removed Ubuntu 14.04 from the list, we should leave
the storage driver choice to every distribution tested.

Furthermore, Fedora does not have the support for AUFS in its kernel,
meaning that this patch is really needed to enable Fedora in our CI.

Fixes #576

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>