metrics: Only systemctl restart once per cycle
Rather than doing a systemctl restart on docker and the proxy
for every test invocation, lets just do it the once at the
start before we begin the test cycle.

Fix up the invocation scripts to make the one off call.

Predominanlty because running many frequent restarts on my
ubuntu machine results in systemd disabling docker.service
as it sees it as restarting too often.

Fixes: #619

Signed-off-by: Graham whaley <graham.whaley@intel.com>