.ci: Split log files to avoid getting too large ones
In case of Jenkins CI, we need to have log files not too large
so that they can be properly moved from the VM to the Jenkins
server as build artifacts. Also, 5 MiB is a reasonable size to
check those logs from a web browser.

Notice that "sudo" has been removed from this script since none
of them were really needed. Moreover, the copy of the runtime
logs with "sudo" was preventing the regular user to modify the
file previously copied.

Fixes #627

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>