metrics: storage: measure cpu and memory as well as bandwidth
Whilst running the storage (fio) tests, also measure the CPU
consuption and the memory footprint being used.
To do this we re-structure the test to set up the initial
test environment (get the container up and then init the
test file - which takes longer than the test itself), and then
we can know exactly when we are running the actual test.

Fixes: #630

Signed-off-by: Graham whaley <graham.whaley@intel.com>