ci: integration: Fix docker kill tests
Recently, the CI has been failing pretty often when running docker
kill tests. This is due to the way the tests are written. Indeed,
because the exit code is retrieved after a time.Sleep(), it does
not ensure the container has properly exited.

This patch inspects State.Status from the container, that way it
makes sure to retrieve the container exit code only after the
container has exited. With this way to proceed, we will never
read an exit code if it is not ready to be read.

Fixes #658

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>