ci: integration: Improve docker kill tests
Docker kill tests rely on "trap" to catch signals and exit with the
specific exit code. Because we need trap to be executed before we
call into "docker kill", the tests wait for 5 seconds to make sure
the trap has been called. This works but that's neither pretty nor
efficient. Instead, let's rely on the logs going out of the container
and try to find a specific tag printed after the trap has been
executed.

Fixes #658

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>