ci: integration: Apply docker timeouts to most tests
We have recently bumped the global timeout from 15s to 60s, and
for all the cases (except docker pull), there is no reason to
apply specific timeouts.

Moreover, this is solving some sporadic issues related to the call
of docker stop. The reason is that docker stop uses SIGTERM first
and after 10s, if the process is still running, it sends SIGHUP.
In both runc and cc-runtime cases, we rely on libcontainer, and
there is a limitation about libcontainer for SIGTERM signal, in
such way that it won't terminate the running process even if the
process receives it.

For those cases, a timeout of 15s for docker stop is too low since
the process has only 5s remaining to really kill and stop everything.

Fixes #664

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>