CI: Jenkins: Use the Jenkins WORKSPACE
Currently we place our GOPATH into our HOME dir.
This will not scale well for Jenkins slaves that are either not
oneshot/transient or run parallel builds.
Utilise the Jenkins provided WORKSPACE area for the gopath, which
also has the added benefit of potentially being cleaned up at the
end of a run.

Fixes: #667

Signed-off-by: Graham Whaley <graham.whaley@intel.com>