metrics: time: Measure 'time to workload', and other relevant times
The 'docker_workload_time.sh' script measures the round-trip time for
a complete container launch/quit. Whilst this is ueful and lets us
take a single measure to spot regressions in the complete cycle, quite
often we'd like to know more fine-grained times, such as how long it
takes to boot to a payload.
Add a test that takes a number of measures through a run/quit cycle,
and then calculates interesting times from that.

The measures taken here are:
 - start time (on host)
 - time at execution of payload (`date`)
 - timestamp of last `dmesg` entry
 - stop time (on host)

From those we calculate:
 - time to get to workload
 - time to get to the VM kernel
 - time to get through VM kernel
 - time to quit
 - total round-trip time

Fixes: #358

Signed-off-by: Graham whaley <graham.whaley@intel.com>