integration: docker: Fix diff test by making sure the container runs
The diff test needs to use -i flag in order to keep the container
process running while a "docker exec" is started. This issue was
not visible before because STDIN was never closed on the agent side,
even if it was on the host side. With recent patch from the agent,
this issue got highlighted.

Fixes #705

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>