functional: Fix exec tests
The functional tests enable terminal=true for exec commands, and at
the same time, they use the old flag --console= without specifying
any path. This leads STDIN to be closed when the command is executed,
and the shim sends this info to the agent, which closes the terminal.
This case should not be testable/acceptable since no terminal is
configured and passed as an argument to the runtime, so that we can
actually use terminal=true for the agent.

This patch forces the use of terminal=false, meaning that STDOUT and
STDERR will still be open to receive the output and it won't cause an
error on the agent.

Fixes #718

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>