ci: metrics: Choose check method according to CI
Depending on which CI system we are running (such as localCI
or Jenkins etc.), we want to use different 'tail end' checks.
Modify the script to check the CI env variables, and do the
appropriate thing.
Note, the 'generic' thing is to use the checkmetrics tool, and
have the config file determined by the name of the host, thus
easily allowing different config files per host (that are not
all called the same thing).

Fixes: #724

Signed-off-by: Graham whaley <graham.whaley@intel.com>