metrics: killall: do not use the binary path in killall
If you have any path characters in the argument to killall
then it will only kill the processes actually using that
binary. Thus, if you have any stuck processes that are using
a different (or previous that has since been replaced) binary,
killall will not find/kill them.
Change the killall to use just the process/file name (with no
path), which means it will kill all processes of that name
(even if they were launched from a binary in a different path)
which may feel a little more 'dangerous', but as we only run
this in reasonably well defined and constrained systems, it
should be fine.

Fixes: #757

Signed-off-by: Graham whaley <graham.whaley@intel.com>