metrics: README: Add a warning that metrics can be destructive
when metrics are run they first try to get the system into a
'known stable state' to reduce the chance of any 'noise injection'
into the results.
If you are unaware of this, this could be potentially destructive and
come as somewhat of a surprise.
Add a warning to the metrics README to try and make it less of a
surprise.

Signed-off-by: Graham whaley <graham.whaley@intel.com>