metrics: latency: use tail instead of a shell for container
We have been using a shell to 'hold open' a container for us to
then work with - but shells are not happy like this when we have a
tty/headless setup (like some CIs or running under nohup for instance).
Switch to using a 'tail -f /dev/null' to hold the container alive.

Fixes: #770

Signed-off-by: Graham whaley <graham.whaley@intel.com>