cmd: add manage_ctr_mgr.sh tool
manage_ctr_mgr.sh is a tool that will help you install and
configure the correct version of docker depending on your
needs. For example, if you need to use Swarm with CC, the
tool will install docker 1.12.1 (which is the currently supported
version). If you do not specify to use swarm, it will install
latest supported version of docker.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>