checkmetrics: Print out how wide the 'acceptance gap' is
When we set the min/max floor/ceiling values for the 'range' that
the result should fall into, it is very hard to tell what 'quality'
that range is - how wide it is compared to the result itself.
Add a column to the results table to show that value (as a %
of the MinVal), so we get a visual clue as to how good or bad
our range settings are.

Fixes: #859

Signed-off-by: Graham Whaley <graham.whaley@intel.com>