checkcommits: allow tests on entire commit
Reworked to retrieve the commit details from `git(1)` before applying checks
to the commit details.

Previously, the "get" and "check" code was somewhat intermingled meaning it
was not possible to apply tests to the overall commit (for example a check
that spanned a commit subject *and* commit body).

Initialised vendoring for assert package.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>