metrics: density: update footprint to cater for kata
Update the footprint script to make it more generic around
proxy and shim paths, and thus able to work with kata as well.
Also fix a niggle/bug where for runc we would have no proxy
or shim, so the uss and pss would report for the whole system
instead of the valid cc components. Now they report 0, which
is more correct.

Signed-off-by: Graham whaley <graham.whaley@intel.com>