CI: Specify full path for clarity in Jenkins setup script
Specify the full path to the repo-specific `setup.sh` script to make it
clear which `setup.sh` script is actually being called. This isn't
essential but does improve the code clarity given that the `pushd` is a
long way from the call site.

Also expanded the comment to further increase clarity.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>