CI: Do not build, but do update, repos when pulling them
Without a '-d' on the 'go get', the script ties to build the repos,
and for repos that do not work under 'go build', this leaves errors
in the CI logs. We don't need a build at this stage as we invoke the
repo specific builds later on.
Also, some CIs do not clean out their build trees between builds, and
thus have repos already present - so add an 'update -u' flag as well.

Fixes: #887

Signed-off-by: Graham Whaley <graham.whaley@intel.com>