ctr_mgr: Remove newgroup command from installation
newgrp command was introduced to login into the docker
group and allow a normal user execute the runtime
unit tests which require the use of docker without
root, but this command didn't had the desired effect
on the CI and when running under a terminal it creates
a new bash session, stopping the installation of all
other components. For this reason, lets remove this
command from the docker installation.

Fixes #890.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>