metrics: storage: add blogbench test
Add blogbench as a test that performs a many-file and process
read and write test.

Build the test from source in the Dockerfile, as that is much easier
to install and run that using the phoronix wrapped version.

We could no doubt form a similar test by configuring `fio`, and if
we need we can examine that later.

Fixes: #880

Signed-off-by: Graham whaley <graham.whaley@intel.com>